Moving tests to a seperate directory
