+++ packages/realm-web/src/tests/utils/MockTransport.ts
@@ -16,7 +16,8 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-import { Transport, PrefixedTransport, BaseTransport } from "../transports";
+import { Transport, PrefixedTransport, BaseTransport } from "../../transports";
+
 import { MockNetworkTransport } from "./MockNetworkTransport";
 
 /**
