+++ packages/realm-web/src/transports/index.ts
@@ -17,7 +17,7 @@
 ////////////////////////////////////////////////////////////////////////////
 
 export { AppTransport } from "./AppTransport";
-export { Transport, Request } from "./Transport";
+export { Transport, Request, MongoDBRealmError } from "./Transport";
 export { BaseTransport } from "./BaseTransport";
 export { AuthenticatedTransport } from "./AuthenticatedTransport";
 export { PrefixedTransport } from "./PrefixedTransport";
