--- packages/realm-web/test/tsconfig.json
@@ -1,6 +1,0 @@
-{
-    "extends": "../tsconfig.json",
-    "compilerOptions": {
-        "noResolve": false
-    }
-}
